ai-teacher/voice-based/requirements.txt
openai-whisper==20231117
anthropic==0.7.8
openai==1.3.8
pyaudio==0.2.11
pydub==0.25.1
numpy==1.24.3
scipy==1.11.4
soundfile==0.12.1
webrtcvad==2.0.10
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
aiofiles==23.2.1
